+++ packages/editor/js/extensions/libs/background/bootstrap.js
@@ -4,6 +4,8 @@
  * External dependencies
  */
 import { addFilter } from '@wordpress/hooks';
+import { createHigherOrderComponent } from '@wordpress/compose';
+import { createElement } from '@wordpress/element';
 
 /**
  * Blockera dependencies
@@ -27,6 +29,10 @@ import {
 	mergeWPCompatibility,
 	sanitizeWPCompatibilityAttributes,
 } from '../utils';
+import {
+	splitConflictingBackgroundStyle,
+	splitConflictingBackgroundWrapperProps,
+} from './compatibility/split-background-style';
 
 export const bootstrap = (): void => {
 	addFilter(
@@ -117,6 +123,41 @@ export const bootstrap = (): void => {
 			}
 
 			return nextState;
+		}
+	);
+
+	addFilter(
+		'editor.BlockListBlock',
+		'blockera.editor.sanitizeBackgroundStyleConflict',
+		createHigherOrderComponent((BlockListBlock) => {
+			return function BlockeraSanitizedBackgroundStyle(props: Object) {
+				return createElement(BlockListBlock, {
+					...props,
+					wrapperProps: splitConflictingBackgroundWrapperProps(
+						props.wrapperProps
+					),
+				});
+			};
+		}, 'withSanitizedBackgroundStyle'),
+		9
+	);
+
+	addFilter(
+		'blocks.getSaveContent.extraProps',
+		'blockera.editor.sanitizeBackgroundStyleConflict.save',
+		(extraProps: Object): Object => {
+			const nextStyle = splitConflictingBackgroundStyle(
+				extraProps?.style
+			);
+
+			if (nextStyle === extraProps?.style) {
+				return extraProps;
+			}
+
+			return {
+				...extraProps,
+				style: nextStyle,
+			};
 		}
 	);
 };